restructure how we do measurements a little
